dom/workers/test/serviceworkers/test_get_serviced.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 982726 - test get_serviced not crashing</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">
  // get_serviced_worker will call getServiced until the worker shuts down.
  // Test passes if the browser doesn't crash on leaked promise objects.
  var controlled_window;
  function simpleRegister() {
    return navigator.serviceWorker.register("get_serviced_worker.js", { scope: "./sw_clients/" });
  }

  function openWindow() {
    var p = new Promise(function(resolve, reject) {
      window.onmessage = function(e) {
        if (e.data === "READY") {
          resolve();
        }
      }
    });

    controlled_window = window.open("sw_clients/simple.html");
    return p;
  }

  function runTest() {
    simpleRegister()
      .then(openWindow).catch(function(e) {
        ok(false, "Some test failed with error " + e);
      }).then(function() {
        ok(true, "Didn't crash on resolving getServiced promises while worker shuts down.");
        SimpleTest.finish();
        controlled_window.close();
      });
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true]
  ]}, runTest);
</script>
</pre>
</body>
</html>
